add: 構成の見直し&Misezan.scaalaのテスㇳテンプレート追加

+++ src/main/scala/MyServer.scala
@@ -16,7 +16,7 @@ object MySever {
       path("misezan") {
         parameters("arg1".as[Int], "arg2".as[Int]) { (arg1, arg2) =>
           val misezan = new Misezan()
-          val result: Int = misezan.calculate(arg1, arg2)
+          val result: Double = misezan.calculate(arg1, arg2)
           complete("眼は" + result.toString + "です\n")
         }
       }
